+++ src/Products/MeetingCharleroi/tests/testWFAdaptations.py
@@ -1,11 +1,10 @@
 # -*- coding: utf-8 -*-
 
-from Products.Archetypes.event import ObjectEditedEvent
+from imio.helpers.content import get_vocab_values
 from Products.MeetingCharleroi.tests.MeetingCharleroiTestCase import MeetingCharleroiTestCase
 from Products.MeetingCharleroi.utils import finance_group_uid
 from Products.MeetingCommunes.tests.testWFAdaptations import testWFAdaptations as mctwfa
 from Products.PloneMeeting.config import MEETING_REMOVE_MOG_WFA
-from zope.event import notify
 
 
 class testWFAdaptations(MeetingCharleroiTestCase, mctwfa):
@@ -15,32 +14,36 @@ class testWFAdaptations(MeetingCharleroiTestCase, mctwfa):
         """Test what are the available wfAdaptations."""
         # we removed the 'archiving' and 'creator_initiated_decisions' wfAdaptations
         self.assertEqual(
-            sorted(self.meetingConfig.listWorkflowAdaptations().keys()),
+            sorted(get_vocab_values(self.meetingConfig, 'WorkflowAdaptations')),
             sorted([
-                "only_creator_may_delete",
-                "no_publication",
-                "accepted_but_modified",
-                "postpone_next_meeting",
-                "itemdecided",
-                "mark_not_applicable",
+                'accepted_but_modified',
+                'charleroi_return_to_any_state_when_prevalidated',
+                'itemdecided',
+                'item_validation_shortcuts',
+                'no_freeze',
+                'no_publication',
+                'no_decide',
+                'only_creator_may_delete',
+                'postpone_next_meeting',
+                'mark_not_applicable',
                 MEETING_REMOVE_MOG_WFA,
-                "removed",
-                "removed_and_duplicated",
-                "refused",
-                "delayed",
-                "pre_accepted",
-                "return_to_proposing_group",
-                "decide_item_when_back_to_meeting_from_returned_to_proposing_group",
-                "hide_decisions_when_under_writing",
-                "waiting_advices",
-                "waiting_advices_proposing_group_send_back",
-                "meetingmanager_correct_closed_meeting",
-                "charleroi_return_to_any_state_when_prevalidated",
-            ]),
+                'removed',
+                'removed_and_duplicated',
+                'refused',
+                'delayed',
+                'pre_accepted',
+                'return_to_proposing_group',
+                'decide_item_when_back_to_meeting_from_returned_to_proposing_group',
+                'hide_decisions_when_under_writing',
+                'waiting_advices',
+                'waiting_advices_proposing_group_send_back',
+                'meetingmanager_correct_closed_meeting',
+            ])
         )
 
     def test_pm_Validate_workflowAdaptations_dependencies(self):
         """Bypass as all WFA are not available"""
+        pass
 
     def test_pm_WFA_pre_validation(self):
         """Will not work as we have also a state before...
@@ -49,49 +52,13 @@ class testWFAdaptations(MeetingCharleroiTestCase, mctwfa):
 
     def test_pm_WFA_waiting_advices_unknown_state(self):
         """Bypass as we have more states from which we can return from waiting advice"""
-
-    def test_pm_WFA_charleroi_return_to_any_state_when_prevalidated(self):
-        """Test that permissions are correct when the WFA is enabled."""
-        cfg = self.meetingConfig
-        self.changeUser("siteadmin")
-        cfg.setWorkflowAdaptations(())
-        notify(ObjectEditedEvent(cfg))
-        itemWF = self.wfTool.getWorkflowsFor(cfg.getItemTypeName())[0]
-        self.assertFalse("backToProposedFromPrevalidated" in itemWF.transitions)
-        self.assertFalse("backToItemCreatedFromPrevalidated" in itemWF.transitions)
-        # activate, needs the 'charleroi_return_to_any_state_when_prevalidated' WFA
-        cfg.setWorkflowAdaptations(("charleroi_return_to_any_state_when_prevalidated",))
-        notify(ObjectEditedEvent(cfg))
-        itemWF = self.wfTool.getWorkflowsFor(cfg.getItemTypeName())[0]
-        self.assertTrue("backToProposedFromPrevalidated" in itemWF.transitions)
-        self.assertTrue("backToItemCreatedFromPrevalidated" in itemWF.transitions)
-
-        # test that transitions do the work
-        self.changeUser("pmManager")
-        item = self.create("MeetingItem")
-
-        # back to proposed to servicehead
-        self.do(item, "propose")
-        self.do(item, "proposeToRefAdmin")
-        self.do(item, "prevalidate")
-        # MeetingMember can not trigger the transitions
-        self.changeUser("pmCreator1")
-        self.assertEqual(self.transitions(item), [])
-        self.changeUser("pmManager")
-        self.do(item, "backToProposedFromPrevalidated")
-        self.assertEqual(item.query_state(), "proposed")
-
-        # back to itemcreated
-        self.do(item, "proposeToRefAdmin")
-        self.do(item, "prevalidate")
-        self.do(item, "backToItemCreatedFromPrevalidated")
-        self.assertEqual(item.query_state(), "itemcreated")
+        pass
 
     def _setItemToWaitingAdvices(self, item, transition=None):
         """We need to ask finances advice to be able to do the transition."""
         originalMember = self.member.getId()
         self.changeUser("siteadmin")
-        self._configureFinancesAdvice(self.meetingConfig)
+        self._configureCharleroiFinancesAdvice(self.meetingConfig)
         self.changeUser(originalMember)
         item.setOptionalAdvisers(
             item.getOptionalAdvisers() + ("{0}__rowid__unique_id_002".format(finance_group_uid()),)
@@ -102,7 +69,7 @@ class testWFAdaptations(MeetingCharleroiTestCase, mctwfa):
 
     def test_pm_WFA_waiting_advices_with_prevalidation(self):
         """Bypass it's tested in testCustomWorkflow."""
-
+        pass
 
     def _userAbleToBackFromWaitingAdvices(self, currentState):
         """Return username able to back from waiting advices."""