+++ CHANGES.rst
@@ -5,8 +5,10 @@ Products.MeetingCharleroi Changelog
 4.2.0rc3 (unreleased)
 ---------------------
 
-- Nothing changed yet.
-
+- Make `no_freeze` and `no_decide` WF adaptations available.
+  [gbastien]
+- Adapted code for `ToolPloneMeeting.advisersConfig`, added upgrade step to `4201`.
+  [gbastien]
 
 4.2.0rc2 (2023-10-20)
 ---------------------